streamlit_website/requirements.txt
# Core Dependencies
streamlit
numpy<2
pandas
plotly
matplotlib
seaborn
joypy
joblib

# Machine Learning & Image Processing
torch
torchvision
torchaudio
opencv-python-headless
pillow

# Serialization & Utilities
dill
pickle-mixin
